+++ minhas-financas-back-end/MinhasFinancas.Application/Interfaces/IMfScoreLaboratorioAppService.cs
@@ -1,5 +1,12 @@
 namespace MinhasFinancas.Application.Interfaces
 {
+    using MinhasFinancas.Application.DTOs.MfScoreLaboratorio;
+
+    public interface IBenchmarkMfScoreService
+    {
+        Task<BenchmarkCenarioMfScoreLaboratorioDTO?> BuscarCenarioAsync(string codigoCenario);
+    }
+
     public interface IMfScoreLaboratorioAppService
     {
         Task<RetornoGenerico> BuscarUsuariosAsync();